+++ src/dlt_filesystem/source/impl/remote.py
@@ -691,3 +691,78 @@ class DropboxSource(FilesystemSource):
                 column_types=kwargs.get("column_types"),
             )
         )
+
+
+class WebdavSource(FilesystemSource):
+    """
+    Access files on WebDAV.
+
+    https://skshetry.github.io/webdav4/
+    https://en.wikipedia.org/wiki/WebDAV
+    """
+
+    @property
+    def fs_class(self) -> Type["AbstractFileSystem"]:
+        from webdav4.fsspec import WebdavFileSystem
+
+        return WebdavFileSystem
+
+    def dlt_source(self, uri: str, table: str, **kwargs):
+
+        # TODO: Is this applicable for Dropbox and friends at all?
+        if kwargs.get("incremental_key"):
+            raise ValueError(
+                "WebDAV takes care of incrementality on its own, you should not provide incremental_key"
+            )
+
+        # Decode URL.
+        parsed_uri = urlparse(uri)
+        parsed_fields = parse_qs(parsed_uri.query)
+
+        # Decode query arguments.
+        fs_kwargs: Dict[str, Any] = {
+            key: value[0] for key, value in parsed_fields.items()
+        }
+        auth = None
+        if "username" in fs_kwargs:
+            auth = (fs_kwargs["username"], fs_kwargs.get("password"))
+
+        # TODO: It looks like this is generic code that could be refactored already
+        #       if it's common amongst different implementations. Breaking out the
+        #       reference to the relevant filesystem implementation itself per
+        #       `fs_class` already contributed to a better situation than before.
+        bucket_name, path_to_file = parse_uri(parsed_uri, table)
+        if not bucket_name or not path_to_file:
+            raise InvalidBlobTableError("WebDAV")
+
+        scheme = parsed_uri.scheme.replace("+webdav", "")
+        bucket_url = f"{scheme}://{bucket_name}/"
+
+        fs = self.fs_class(uri, auth=auth)
+
+        # TODO: Naming things: Rename `determine_endpoint` to `find_reader`.
+        # TODO: Refactoring: Break out reader finding and fragments of the
+        #       filesystem initialization into common routines.
+        try:
+            endpoint: str = determine_endpoint(table, path_to_file)
+        except UnsupportedEndpointError:
+            raise ValueError(supported_file_format_message("WebDAV"))
+        except Exception as e:
+            raise ValueError(
+                f"Failed to parse endpoint from path: {path_to_file}"
+            ) from e
+
+        from dlt_filesystem.source.adapter import resource_for_reader
+        from dlt_filesystem.source.model import FilesystemReference
+
+        return resource_for_reader(
+            FilesystemReference(
+                fs=fs,
+                bucket_url=bucket_url,
+                file_glob=path_to_file,
+                reader_name=endpoint,
+                hints=blob_hints(parsed_uri, table),
+                # TODO: Can `column_types` be looped into reader hints instead?
+                column_types=kwargs.get("column_types"),
+            )
+        )