+++ tests/main/filesystem/test_remote_read.py
@@ -51,6 +51,7 @@ URIS = [
         uri="hdfs://example.com:8020?user=test",
         table="path/to/data.parquet",
     ),
+    "http+webdav://username:password@www.example.org/path/to/data.parquet",
     f'oci://bucket@namespace/prefix/path/to/data.parquet?iam_type=api_key&config={{"user":"ocid1.user.oc1..24g4uzg","region":"us-ashburn-1","tenancy":"ocid1.tenancy.oc1..23423r3","key_file":"{private_key_file}","fingerprint":"{private_key_fingerprint}"}}',
     "oss://bucket/path/to/data.parquet?endpoint=http://oss-cn-hangzhou.aliyuncs.com/&key=foo&secret=bar",
     "oss://bucket/path/to/data.parquet?endpoint=https://oss-me-east-1.aliyuncs.com/&token=foobar",